не помню

+++ MinesServer/GameShit/GUI/Horb/CrysLine.cs
@@ -1,4 +1,4 @@
 ﻿namespace MinesServer.GameShit.GUI.Horb
 {
-    public readonly record struct CrysLine(string Label, int LeftMin, int RightMin, long Denominator, int CurrentValue) { }
+    public readonly record struct CrysLine(string Label, long LeftMin, long RightMin, long Denominator, long CurrentValue) { }
 }
