+++ MinesServer/GameShit/Sys_Market/MarketSystem.cs
@@ -54,12 +54,12 @@ namespace MinesServer.GameShit.SysMarket
             {
                 OnAdmin = (p.id != m.ownerid ? null : () => m.onadmn(p, m)),
                 CrystalConfig = new CrystalConfig(" ", "цена", [
-                            new CrysLine($"<color=#aaeeaa>{World.GetCrysCost(0) * 10}$</color>", 0, 0, (int)(p.money / (World.GetCrysCost(0) * 10)), 0),
-                    new CrysLine($"<color=#aaeeaa>{World.GetCrysCost(1) * 10}$</color>", 0, 0, (int)(p.money / (World.GetCrysCost(1) * 10)), 0),
-                    new CrysLine($"<color=#aaeeaa>{World.GetCrysCost(2) * 10}$</color>", 0, 0, (int)(p.money / (World.GetCrysCost(2) * 10)), 0),
-                    new CrysLine($"<color=#aaeeaa>{World.GetCrysCost(3) * 10}$</color>", 0, 0, (int)(p.money / (World.GetCrysCost(3) * 10)), 0),
-                    new CrysLine($"<color=#aaeeaa>{World.GetCrysCost(4) * 10}$</color>", 0, 0, (int)(p.money / (World.GetCrysCost(4) * 10)), 0),
-                    new CrysLine($"<color=#aaeeaa>{World.GetCrysCost(5) * 10}$</color>", 0, 0, (int)(p.money / (World.GetCrysCost(5) * 10)), 0)
+                            new CrysLine($"<color=#aaeeaa>{World.GetCrysCost(0) * 10}$</color>", 0, 0, (p.money / (World.GetCrysCost(0) * 10)), 0),
+                    new CrysLine($"<color=#aaeeaa>{World.GetCrysCost(1) * 10}$</color>", 0, 0, (p.money / (World.GetCrysCost(1) * 10)), 0),
+                    new CrysLine($"<color=#aaeeaa>{World.GetCrysCost(2) * 10}$</color>", 0, 0, (p.money / (World.GetCrysCost(2) * 10)), 0),
+                    new CrysLine($"<color=#aaeeaa>{World.GetCrysCost(3) * 10}$</color>", 0, 0, (p.money / (World.GetCrysCost(3) * 10)), 0),
+                    new CrysLine($"<color=#aaeeaa>{World.GetCrysCost(4) * 10}$</color>", 0, 0, (p.money / (World.GetCrysCost(4) * 10)), 0),
+                    new CrysLine($"<color=#aaeeaa>{World.GetCrysCost(5) * 10}$</color>", 0, 0, (p.money / (World.GetCrysCost(5) * 10)), 0)
 
                             ], true),
                 Text = $"Покупка\nКупленно кристалов на <color=#aaeeaa>{-money}$</color>",
